fix: GH login
